==> mahdi/2.Clean_Numeric_Data.ipynb <==
# 2. Clean Numeric Data

## Things to consider:
    - [ ] Range of the variables
    - [ ] Nulls in each feature column. Remove columns with number of nulls bigger than a threshold.
    - [ ] One value dominance in each feature column. Remove columns with dominance bigger than a threshold.
    - [ ] Remove rows that have nulls
    - [ ] Cleaning the test dataset?

import numpy as np 
import matplotlib.pyplot as plt
from scipy import stats
import pandas as pd
%matplotlib inline

## Load data

train_numeric_file = 'house-prices-advanced-regression-techniques/train_numeric.csv'
test_numeric_file = 'house-prices-advanced-regression-techniques/test_numeric.csv'

## Cleaning the training data

train_df = pd.read_csv(train_numeric_file)
print(train_df.head())

### Check th range of the features
    - NOT IMPLEMENTED 

### Remove columns with number of nulls bigger than a threshold.  
### Remove columns with dominance bigger than a threshold.

drop_thresh = 0.8
drop_list = []
for feature in train_df:
    col_df = train_df[feature]
    count_nan = col_df.isnull().sum()
    nan_ratio = count_nan/len(col_df)   
    repeats = train_df.pivot_table(index=[feature], aggfunc='size').sort_values()
    max_repeat_ratio = repeats.max()/len(col_df)
    if (nan_ratio>drop_thresh or max_repeat_ratio>drop_thresh):
        drop_list.append(feature)
        
train_df_repeats_removed = train_df.drop(drop_list, axis=1)

print(train_df.shape)
print(train_df_repeats_removed.shape)
    
    

### Remove rows that contain null

train_df_repeats_and_nulls_removed = train_df_repeats_removed.dropna()
print(train_df_repeats_and_nulls_removed.shape)

### Save the clean train data

train_clean = train_df_repeats_and_nulls_removed
train_clean.to_csv('house-prices-advanced-regression-techniques/train_clean.csv', index=False)

## Cleaning the test data

### Drop the non-valid cols

test_df = pd.read_csv(test_numeric_file)
print(test_df.shape)
test_df_valid_cols = test_df.drop(drop_list, axis=1)
print(test_df_valid_cols.shape)


### Replace null values with average of the column. We do not want to loose any test row.

CAUTION: if you want to use this for Machine Learning / Data Science: from a Data Science perspective it is wrong to first replace NA and then split into train and test... You MUST first split into train and test, then replace NA by mean on train and then apply this stateful preprocessing model to test,

for feature in test_df_valid_cols:
    test_df_valid_cols[feature].fillna((train_clean[feature].mean()), inplace=True)
    
    
test_df_valid_cols_nulls_replaced = test_df_valid_cols
print(test_df_valid_cols_nulls_replaced.shape)

### Save the clean test data

test_clean = test_df_valid_cols_nulls_replaced
test_clean.to_csv('house-prices-advanced-regression-techniques/test_clean.csv', index=False)
