==> mahdi/3.Outlier_Detection.ipynb <==
# 3. Outlier Detection

## A good source to start with:
https://towardsdatascience.com/5-ways-to-detect-outliers-that-every-data-scientist-should-know-python-code-70a54335a623
    

import numpy as np 
import matplotlib.pyplot as plt
from scipy import stats
import pandas as pd
%matplotlib inline

## Load data

train_file = 'house-prices-advanced-regression-techniques/train_clean.csv'
train_df = pd.read_csv(train_file)
print(train_df.head())


## Method 1,  Standard Deviation: Good for single dimension outlier detection

In statistics, If a data distribution is approximately normal then about 68% of the data values lie within one standard deviation of the mean and about 95% are within two standard deviations, and about 99.7% lie within three standard deviations.
Therefore, if you have any data point that is more than 3 times the standard deviation, then those points are very likely to be anomalous or outliers.

import numpy as np
import matplotlib.pyplot as plt


# Function to Detection Outlier on one-dimentional datasets.
def find_anomalies(data):
    anomalies_idx = []
    # Set upper and lower limit to 3 standard deviation
    data_std = data.std()
    data_mean = data.mean()
    anomaly_cut_off = data_std * 3
    
    lower_limit  = data_mean - anomaly_cut_off 
    upper_limit = data_mean + anomaly_cut_off
    # Generate outliers
    for idx in range(len(data)):
        outlier = data[idx]
        if outlier > upper_limit or outlier < lower_limit:
            anomalies_idx.append(idx)
    return anomalies_idx

outliers_full_list = []
for feature in train_df:
    outliers_full_list = outliers_full_list + find_anomalies(train_df[feature].values)

outliers_unique = list(set(outliers_full_list))
print(len(outliers_full_list))
print(len(outliers_unique))

train_df_outlier_removed = train_df.drop(outliers_unique)

print(train_df.shape)
print(train_df_outlier_removed.shape)

### Save the outlier removed training data

train_df_outlier_removed.to_csv('house-prices-advanced-regression-techniques/train_outlier_removed.csv', index=False)
