==> mahdi/6.Dimensionality_Reduction.ipynb <==
# 6. Dimensionality Reduction
I am following this interesting instruction for PCA: https://towardsdatascience.com/pca-using-python-scikit-learn-e653f8989e60

import numpy as np 
import matplotlib.pyplot as plt
from scipy import stats
import pandas as pd
%matplotlib inline

train_file = 'house-prices-advanced-regression-techniques/train_normalized.csv'
train_df = pd.read_csv(train_file)
test_file = 'house-prices-advanced-regression-techniques/test_normalized.csv'
test_df = pd.read_csv(test_file)

label_col = train_df['SalePrice']
data_df = train_df.drop(['SalePrice'], axis=1)
print(train_df.head())

## Import and Apply PCA
Notice the code below has .95 for the number of components parameter. It means that scikit-learn choose the minimum number of principal components such that 95% of the variance is retained.

from sklearn.decomposition import PCA
# Make an instance of the Model
pca = PCA(.95)

Fit PCA on training set. Note: you are fitting PCA on the training set only.

pca.fit(data_df)
print(pca.n_components_)

Note: You can find out how many components PCA choose after fitting the model using pca.n_components_ . In this case, 95% of the variance amounts to 40 principal components.

train_reduced = pca.transform(data_df)
test_reduced = pca.transform(test_df)

new_features_list = ['PCA_Feature_'+str(x) for x in range(pca.n_components_)]

train_reduced_df = pd.DataFrame(train_reduced, columns=new_features_list)
test_reduced_df = pd.DataFrame(test_reduced, columns=new_features_list)
print(train_reduced_df.shape)
print(test_reduced_df.shape)

## Save the dimension reduced train and test data

train_reduced_with_label = pd.concat([train_reduced_df, label_col], axis=1)
train_reduced_with_label.to_csv('house-prices-advanced-regression-techniques/train_dim_reduced.csv', index=False)
test_reduced_df.to_csv('house-prices-advanced-regression-techniques/test_dim_reduced.csv', index=False)
