==> mahdi/11.Merge_ANN_&_RF.ipynb <==
# 11. Merge Artificial Neural Network & Random Forest

import pandas as pd

ann_file = 'Results/ANN_reg_WO_PCA.csv'
rf_file = 'Results/RF_reg_WO_PCA.csv'

ann_DF = pd.read_csv(ann_file)
rf_DF = pd.read_csv(rf_file)

ann_sale = ann_DF['SalePrice']
rf_sale = rf_DF['SalePrice']

output_sale = pd.concat([ann_sale, rf_sale], axis=1).mean(axis=1).round(1)
output_ID = ann_DF['Id']

print(output_sale.head())

## Saving the results

output_DF = pd.concat([output_ID, output_sale], axis=1)
output_DF.columns = ['Id', 'SalePrice']
output_DF.to_csv('ANN_&_RF_WO_PCA.csv', index=False)
